test(rest-api): test the enum-trait for ReleaseAdders, not in the github file
The test covers the github case as well, and this allows us to get rid
of the new_with_client method that was used in the test there.
