+++ rest-api/src/file_auth/github_release.rs
@@ -196,25 +196,6 @@ impl<C: GithubClientTrait> GithubReleaseAdder<C> {
         &self.release_info
     }
 
-    pub async fn new_with_client(
-        release_url: &url::Url,
-        git_repo_path: PathBuf,
-        client: C,
-    ) -> Result<Self, crate::file_auth::release_types::ReleaseUrlError> {
-        let release_info = parse_release_url(release_url, &git_repo_path)
-            .await
-            .map_err(|e| {
-                crate::file_auth::release_types::ReleaseUrlError::InvalidFormat(e.to_string())
-            })?;
-
-        Ok(Self {
-            release_url: release_url.clone(),
-            git_repo_path,
-            client,
-            release_info,
-        })
-    }
-
     fn extract_assets(&self, release: &Release) -> Vec<ReleaseAssetInfo> {
         release
             .assets
@@ -421,8 +402,6 @@ pub mod test_utils {
 
 #[cfg(test)]
 mod tests {
-    #[cfg(feature = "test-utils")]
-    use super::test_utils::MockGithubClient;
     use super::*;
     use tempfile::TempDir;
 
@@ -461,61 +440,5 @@ mod tests {
         let result = parse_release_url(&url, &git_repo).await;
 
         assert!(result.is_err());
-    }
-
-    #[tokio::test]
-    #[cfg(feature = "test-utils")]
-    async fn test_github_release_with_mock() {
-        use super::test_utils::create_mock_release;
-        use crate::file_auth::release_types::ReleaseAdder;
-
-        let temp_dir = TempDir::new().unwrap();
-        let git_repo_path = temp_dir.path().to_path_buf();
-
-        let signers_dir = git_repo_path
-            .join("github.com/testowner/testrepo")
-            .join(SIGNERS_DIR);
-        fs::create_dir_all(&signers_dir).await.unwrap();
-
-        let signers_json = r#"{
-            "version": 1,
-            "required_signers": 1,
-            "signers": [
-                {
-                    "public_key": "test_key",
-                    "name": "Test Signer"
-                }
-            ]
-        }"#;
-        fs::write(signers_dir.join(SIGNERS_FILE), signers_json)
-            .await
-            .unwrap();
-
-        let release_url =
-            url::Url::parse("https://github.com/testowner/testrepo/releases/tag/v1.0.0").unwrap();
-
-        let mock_client = MockGithubClient::new();
-
-        let mut adder =
-            GithubReleaseAdder::new_with_client(&release_url, git_repo_path.clone(), mock_client)
-                .await
-                .unwrap();
-
-        let mock_release = create_mock_release();
-        adder.client.mock_release(mock_release);
-
-        let index_content = adder.index_content().await.unwrap();
-        let json: serde_json::Value = serde_json::from_str(&index_content).unwrap();
-
-        assert_eq!(json["version"], 1);
-        assert!(json["publishedFiles"].is_array());
-        let files = json["publishedFiles"].as_array().unwrap();
-        assert_eq!(files.len(), 1);
-        assert_eq!(files[0]["fileName"], "test.tar.gz");
-        assert_eq!(files[0]["algo"], "Sha256");
-        assert_eq!(
-            files[0]["hash"],
-            "abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
-        );
     }
 }
